rolled back some changes in the code since it needs a bit more thought and refactoring
