--- client/src/app/shared/interfaces/recipe_data.ts
@@ -1,4 +1,0 @@
-export interface RecipeData {
-  id: number;
-  playlistId: string;
-}
