--- client/src/app/shared/interfaces/recipe_response.ts
@@ -1,6 +1,0 @@
-import { Recipe } from './recipe';
-
-export interface RecipeResponse {
-  recipe: Recipe;
-  playlist: string;
-}
