--- client/src/app/shared/interfaces/recipe_scrapped.ts
@@ -1,4 +1,0 @@
-export interface RecipeScrapped {
-  genre: string,
-  url: string
-}
